// source/AdventureWorksAIWorkspaceGUI/src/api/generated/model/updateUserCommand.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * AdventureWorksAIWorkspaceAPI.Api
 * OpenAPI spec version: 1.0
 */

export interface UpdateUserCommand {
  userId: string
  /** @nullable */
  userName?: string | null
  /** @nullable */
  email?: string | null
  /** @nullable */
  role?: string | null
  resetPassword: boolean
}